refactor : common 폴더 shared로 이름 변경
