+++ src/auth/application/usecases/logout.usecase.ts
@@ -1,5 +1,5 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { AuthPlatform } from 'src/common/types/auth-platform.type';
+import { AuthPlatform } from 'src/shared/types/auth-platform.type';
 import { AUTH_SESSION_PORT } from '../ports/auth-session.port';
 import type { AuthSessionPort } from '../ports/auth-session.port';
 import { LogoutOutput } from '../dtos/logout-output.dto';
