+++ src/auth/infrastructure/prisma-auth.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { AuthProvider } from 'src/common/types/auth-provider.type';
+import { AuthProvider } from 'src/shared/types/auth-provider.type';
 import {
   AuthProvider as PrismaAuthProvider,
   SubscriptionPlan,
