+++ src/auth/presentation/strategies/google.strategy.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
-import { AuthProvider } from 'src/common/types/auth-provider.type';
+import { AuthProvider } from 'src/shared/types/auth-provider.type';
 import { Request } from 'express';
 import type { AuthenticateOptions } from 'passport';
 import { Profile, Strategy } from 'passport-google-oauth20';
