+++ src/clips/application/errors/clips.error.ts
@@ -1,7 +1,7 @@
 import {
   ApplicationError,
   ApplicationErrorCode,
-} from '../../../common/application/application.error';
+} from '../../../shared/application/application.error';
 
 export type ClipsErrorCode = Extract<
   ApplicationErrorCode,
