+++ src/shared/application/application.error.ts
