+++ src/shared/presentation/dtos/error-response.dto.ts
