+++ src/shared/types/auth-platform.type.ts
