+++ src/shared/types/auth-provider.type.ts
