+++ src/users/presentation/users.controller.ts
@@ -16,9 +16,9 @@ import {
   ApiTags,
   ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
-import { JwtAccessGuard } from 'src/common/presentation/guards/jwt-access.guard';
-import { AuthContext } from 'src/common/types/auth-context.type';
-import { ApplicationExceptionFilter } from 'src/common/presentation/filters/application-exception.filter';
+import { JwtAccessGuard } from 'src/shared/presentation/guards/jwt-access.guard';
+import { AuthContext } from 'src/shared/types/auth-context.type';
+import { ApplicationExceptionFilter } from 'src/shared/presentation/filters/application-exception.filter';
 import { GetMeUseCase } from '../application/usecases/get-me.usecase';
 import { UpdateMeUseCase } from '../application/usecases/update-me.usecase';
 import { DeleteMeUseCase } from '../application/usecases/delete-me.usecase';
@@ -31,7 +31,7 @@ import {
   UserProfileResponseDto,
   UserSettingsResponseDto,
 } from './dtos/user-response.dto';
-import { ErrorResponseDto } from 'src/common/presentation/dtos/error-response.dto';
+import { ErrorResponseDto } from 'src/shared/presentation/dtos/error-response.dto';
 
 @Controller('users')
 @UseFilters(ApplicationExceptionFilter)
